backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
chromadb
pypdf
numpy
tenacity
cachetools
langgraph
langchain-core
google-cloud-aiplatform
vertexai
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
pytest
httpx
